Overrode getValidRules() and evaluateRule() - moved from Driver.py to DriverEV.py - checks for and returns valid RSev rules as well as RS and RSint rules
